Minor fixes to CombatGame and server packages

+++ src/main/java/com/halenteck/server/Server.java
@@ -324,7 +324,7 @@ public final class Server {
             }
             out.writeByte(CREATE_LOBBY);
             out.writeUTF(lobbyName);
-            int lobbyId = in.readByte();
+            int lobbyId = in.readInt();
             joinLobby(playerName, lobbyId);
         } catch (Exception e) {
             e.printStackTrace();
